Made xcube.core.zarrstore an own package
