+++ test/core/test_xarray.py
@@ -1,13 +1,11 @@
 import unittest
 
-import pytest
 import xarray as xr
 
 from test.sampledata import new_test_dataset
 from xcube.core.gridmapping import GridMapping
 from xcube.core.new import new_cube
 from xcube.core.xarray import DatasetAccessor
-from xcube.core.store.zarrstore import GenericZarrStore
 
 
 class DatasetAccessorTest(unittest.TestCase):
@@ -55,20 +53,25 @@ class DatasetAccessorTest(unittest.TestCase):
         self.assertIsInstance(xr.Dataset.xcube.new(), xr.Dataset)
 
     def test_select_vars(self):
-        self.assertIsInstance(xr.Dataset().xcube.select_variables_subset(), xr.Dataset)
+        self.assertIsInstance(xr.Dataset().xcube.select_variables_subset(),
+                              xr.Dataset)
 
     def test_dump(self):
         self.assertIsInstance(xr.Dataset().xcube.dump(), str)
 
     def test_vars_to_dim(self):
-        dataset = new_test_dataset(["2010-01-01", "2010-01-02", "2010-01-03", "2010-01-04", "2010-01-05"],
-                                   precipitation=0.4, temperature=275.2)
+        dataset = new_test_dataset(
+            ["2010-01-01", "2010-01-02", "2010-01-03", "2010-01-04",
+             "2010-01-05"],
+            precipitation=0.4, temperature=275.2)
 
         self.assertIsInstance(dataset.xcube.vars_to_dim(), xr.Dataset)
 
     def test_levels(self):
-        dataset = new_test_dataset(["2010-01-01", "2010-01-02", "2010-01-03", "2010-01-04", "2010-01-05"],
-                                   precipitation=0.4, temperature=275.2)
+        dataset = new_test_dataset(
+            ["2010-01-01", "2010-01-02", "2010-01-03", "2010-01-04",
+             "2010-01-05"],
+            precipitation=0.4, temperature=275.2)
         levels = dataset.xcube.levels(spatial_tile_shape=(45, 45))
         self.assertIsInstance(levels, list)
         self.assertEqual(3, len(levels))
@@ -77,7 +80,8 @@ class DatasetAccessorTest(unittest.TestCase):
         self.assertTrue(all("temperature" in level for level in levels))
         self.assertEqual([(5, 180, 360), (5, 90, 180), (5, 45, 90)],
                          [level.precipitation.shape for level in levels])
-        self.assertEqual([((1, 1, 1, 1, 1), (45, 45, 45, 45), (45, 45, 45, 45, 45, 45, 45, 45)),
+        self.assertEqual([((1, 1, 1, 1, 1), (45, 45, 45, 45),
+                           (45, 45, 45, 45, 45, 45, 45, 45)),
                           ((1, 1, 1, 1, 1), (45, 45), (45, 45, 45, 45)),
                           ((1, 1, 1, 1, 1), (45,), (45, 45))],
                          [level.precipitation.chunks for level in levels])
